Add files via upload
This shows that I was able to convert a PDF file to a Spreadsheet in Excel. Its not much different than trying to do it in Google Sheets. I then created my own table of data to work with. Preserving the original data. I used Power Query to adjust some of the column data so that it would only give me the Year in column A, and converted the rest of the numbers to the number format. Both were assigned the General format. I will use this data for another project in the future.